Refactor: MainHeader, MainPage 로고 image, title 변경
