+++ src/components/common/Header/MainHeader.tsx
@@ -1,16 +1,16 @@
 import Image from 'next/image';
 import Link from 'next/link';
-import headerIcon from '@/src/assets/icons/headerLogo.svg';
-import headerTitle from '@/src/assets/icons/headerTitle.svg';
 import styles from './MainHeader.module.scss';
+import headerIcon from '@/public/Logo.png';
+import headerTitle from '@/public/Title.png';
 
 const MainHeader = () => {
   return (
     <>
       <header className={styles.mainHeader}>
         <Link href={'/'} className={styles.headerLogo}>
-          <Image src={headerIcon} alt="logo image" />
-          <Image src={headerTitle} alt="logo title" />
+          <Image src={headerIcon} alt="logo image" width={50} height={40} layout="intrinsic" />
+          <Image src={headerTitle} alt="logo title" width={120} height={40} layout="intrinsic" />
         </Link>
         <div className={styles.linkBox}>
           <Link href={'/Login'} className={styles.signLink}>
