+++ src/pages/index.module.scss
@@ -1,5 +1,6 @@
 .homePageLayout {
   display: flex;
+  box-sizing: border-box;
   flex-direction: column;
   align-items: center;
   justify-content: center;
@@ -7,15 +8,16 @@
 
 .mainSection {
   display: flex;
+  width: 120rem;
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  margin: 9.4rem 0;
+  margin: 9rem 0;
 }
 
 .coverImg {
-  width: 72.2rem;
-  height: 42.3rem;
+  // width: 72.2rem;
+  // height: 42.3rem;
   flex-shrink: 0;
   border-radius: 8px;
 }
@@ -37,9 +39,9 @@
   text-align: center;
 }
 
-%mainArticleBox {
+.mainArticleBox {
   display: flex;
-  width: 120rem;
+  width: 100%;
   height: 60rem;
   flex-shrink: 0;
   align-items: center;
@@ -47,7 +49,7 @@
   background: $violet-F1;
 }
 
-%mainArticleImg {
+.mainArticleImg {
   display: flex;
   width: 59.4rem;
   height: 49.8rem;
@@ -64,32 +66,39 @@
   justify-content: center;
   margin-bottom: 10rem;
   gap: 10rem;
+}
 
-  .mainArticleBox:nth-child(odd) {
-    @extend %mainArticleBox;
+@media screen and (width >= 1201px) {
+  .mainArticleBox {
+    &:nth-child(odd) {
+      flex-direction: row;
+      padding-left: 10rem;
 
-    flex-direction: row;
-    padding-left: 10rem;
+      .mainArticleImg {
+        margin-left: auto;
+      }
+    }
 
-    .mainArticleImg {
-      @extend %mainArticleImg;
+    &:nth-child(even) {
+      flex-direction: row-reverse;
+      padding-right: 10rem;
 
-      margin-left: auto;
+      .mainArticleImg {
+        margin-right: auto;
+      }
     }
   }
+}
 
-  .mainArticleBox:nth-child(even) {
-    @extend %mainArticleBox;
-
-    flex-direction: row-reverse;
-    padding-right: 10rem;
-
-    .mainArticleImg {
-      @extend %mainArticleImg;
-
-      margin-right: auto;
-    }
+@media screen and (width <= 1200px) {
+  .mainArticleBox {
+    flex-direction: column;
   }
+}
+
+.logoTitle {
+  width: 400px;
+  height: 120px;
 }
 
 .mainArticleTextBox {
